+++ src/SME.SGP.WebClient/src/servicos/Paginas/ConselhoClasse/ServicoConselhoClasse.js
@@ -1,5 +1,7 @@
+import { conselhoClasseRecomendacaoTipo } from '~/dtos';
 import { store } from '~/redux';
 import {
+  setListaoRecomendacoesAlunoFamilia,
   setListaTiposConceitos,
   setMarcadorParecerConclusivo,
 } from '~/redux/modulos/conselhoClasse/actions';
@@ -201,8 +203,34 @@ class ServicoConselhoClasse {
     return null;
   };
 
-  obterListaAnotacoesRecomendacoes = () =>
-    api.get('/v1/conselhos-classe/obter-recomendacoes');
+  obterListaAnotacoesRecomendacoes = async () => {
+    const state = store.getState();
+
+    const { dispatch } = store;
+    const { conselhoClasse } = state;
+    const listaoRecomendacoesAlunoFamilia =
+      conselhoClasse?.listaoRecomendacoesAlunoFamilia;
+
+    const aluno = listaoRecomendacoesAlunoFamilia?.listaRecomendacoesAluno;
+    const familia = listaoRecomendacoesAlunoFamilia?.listaRecomendacoesFamilia;
+    if (!aluno?.length || !familia?.length) {
+      const retorno = await api.get('/v1/conselhos-classe/obter-recomendacoes');
+
+      const recomendacoes = retorno?.data;
+      if (recomendacoes.length) {
+        const params = {
+          listaRecomendacoesAluno: recomendacoes.filter(
+            item => item.tipo === conselhoClasseRecomendacaoTipo.Aluno
+          ),
+          listaRecomendacoesFamilia: recomendacoes.filter(
+            item => item.tipo === conselhoClasseRecomendacaoTipo.Familia
+          ),
+        };
+
+        dispatch(setListaoRecomendacoesAlunoFamilia(params));
+      }
+    }
+  };
 }
 
 export default new ServicoConselhoClasse();
